refac: Refactored many-to-many for Tour and Equipment
- Added some interfaces and refactored DIP
- Removed https settings for Kestrel
- Added REST endpoint in TourController for adding new Equipment

+++ src/Modules/Tours/Explorer.Tours.Core/Domain/RepositoryInterfaces/ITourRepository.cs
@@ -5,5 +5,5 @@ namespace Explorer.Tours.Core.Domain.RepositoryInterfaces;
 
 public interface ITourRepository : ICrudRepository<Tour>
 {
-    Result UpdateTourEquipment(int tourId, List<int> equipmentIds);
+    Result UpdateTourEquipment(long tourId, List<long> equipmentIds);
 }
